refactor(dev): Adding a storage plugin entrypoint. It will allow to have other plugins to handle notification and log formating.
