+++ cyclosible/playbook/plugins/storage/base.py
